Show state legend for packages with multiple states
