+++ packages/admin-ui/src/pages/packages/components/PackagesList.tsx
@@ -4,7 +4,7 @@ import { useApi } from "api";
 import { packageRestart } from "../actions";
 // Components
 import { NoPackagesYet } from "./NoPackagesYet";
-import { StateBadgeDnp } from "./StateBadge";
+import { StateBadgeDnp, StateBadgeLegend } from "./StateBadge";
 import Card from "components/Card";
 // Icons
 import { MdRefresh, MdOpenInNew } from "react-icons/md";
@@ -35,42 +35,48 @@ export const PackagesList = ({ coreDnps }: { coreDnps: boolean }) => {
       if (!filteredDnps.length) return <NoPackagesYet />;
 
       return (
-        <Card className="list-grid dnps no-a-style">
-          <header className="center">Status</header>
-          <header className="center"> </header>
-          <header>Name</header>
-          <header>Open</header>
-          <header className="hide-on-small">Restart</header>
-          {sortBy(filteredDnps, dnp => dnp.dnpName).map(dnp => (
-            <React.Fragment key={dnp.dnpName}>
-              {/* <StateBadge state={getWorstState(dnp)} /> */}
-              <StateBadgeDnp dnp={dnp} />
-              <img
-                className="avatar"
-                src={dnp.avatarUrl || (coreDnps ? dappnodeIcon : defaultAvatar)}
-                // Display the broken image logo with no text
-                alt=" "
-              />
-              <NavLink
-                className="name"
-                to={urlJoin(packagesRootPath, dnp.dnpName)}
-              >
-                {shortNameCapitalized(dnp.dnpName)}
-              </NavLink>
-              <NavLink
-                className="open"
-                to={urlJoin(packagesRootPath, dnp.dnpName)}
-              >
-                <MdOpenInNew />
-              </NavLink>
-              <MdRefresh
-                className="hide-on-small"
-                style={{ fontSize: "1.05rem" }}
-                onClick={() => packageRestart(dnp).catch(console.error)}
-              />
-              <hr />
-            </React.Fragment>
-          ))}
+        <Card spacing>
+          <StateBadgeLegend dnps={filteredDnps}></StateBadgeLegend>
+
+          <div className="list-grid dnps no-a-style">
+            <header className="center">Status</header>
+            <header className="center"> </header>
+            <header>Name</header>
+            <header>Open</header>
+            <header className="hide-on-small">Restart</header>
+            {sortBy(filteredDnps, dnp => dnp.dnpName).map(dnp => (
+              <React.Fragment key={dnp.dnpName}>
+                {/* <StateBadge state={getWorstState(dnp)} /> */}
+                <StateBadgeDnp dnp={dnp} />
+                <img
+                  className="avatar"
+                  src={
+                    dnp.avatarUrl || (coreDnps ? dappnodeIcon : defaultAvatar)
+                  }
+                  // Display the broken image logo with no text
+                  alt=" "
+                />
+                <NavLink
+                  className="name"
+                  to={urlJoin(packagesRootPath, dnp.dnpName)}
+                >
+                  {shortNameCapitalized(dnp.dnpName)}
+                </NavLink>
+                <NavLink
+                  className="open"
+                  to={urlJoin(packagesRootPath, dnp.dnpName)}
+                >
+                  <MdOpenInNew />
+                </NavLink>
+                <MdRefresh
+                  className="hide-on-small"
+                  style={{ fontSize: "1.05rem" }}
+                  onClick={() => packageRestart(dnp).catch(console.error)}
+                />
+                <hr />
+              </React.Fragment>
+            ))}
+          </div>
         </Card>
       );
     }
